intro-to-ai/project-2/requirements.txt
torch==2.2.1
torchvision==0.17.1
datasets==2.14.2
transformers==4.57.1
Pillow==9.5.0
numpy<2
tqdm==4.67.1
